python_wrapper/requirements.txt
# Python Serial Data Collector Dependencies
pyserial>=3.5

# Azure IoT Hub integration (optional)
azure-iot-device>=2.12.0